requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.2
filelock==3.12.2
h11==0.14.0
idna==3.4
outcome==1.2.0
platformdirs==3.9.1
psycopg2-binary==2.9.6
PySocks==1.7.1
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2023.3
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
wsproto==1.2.0
